Make forms use the same component
